Fixed keyPath bug
Typo with keyPath vs keyPaths
objectStore.put takes two arguments - value and key.
If you already specified the keyPath, no need to specify key.
Otherwise, key takes a value that serves as a unique ID, NOT the name
of one of the object's properties.
For example, key could be "falsjdf5509s", but not "uuid"